content/base/test/test_bug28293.xhtml

<html xmlns="http://www.w3.org/1999/xhtml">
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=28293
-->
<head>
  <title>Test for Bug 28293</title>
  <script type="text/javascript" src="/MochiKit/packed.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
  <script>
res = 'A';

SimpleTest.waitForExplicitFinish();
onload = function () {

  res+='2';

  s = document.createElement('script');
  s.textContent="res+='g';";
  s.defer = true;
  document.body.appendChild(s);

  res+='3';

  s = document.createElement('script');
  s.src="file_bug28293.sjs?res+='h';";
  s.defer = true;
  document.body.appendChild(s);

  s = document.createElement('script');
  s.textContent="res+='i';done()";
  s.defer = true;
  document.body.appendChild(s);

  res+='4';
}

function done() {
  is(res, "AacBCDEFGeHIJfbd1M2g34hi", "scripts executed in the wrong order");
  ok(!fHadExecuted, "Dynamic script executed too late");
  SimpleTest.finish();
}
</script>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=28293">Mozilla Bug 28293</a>

<script defer="defer">
res += 'a';
</script>
<script defer="defer" src="data:text/plain,res+='b'"></script>
<script defer="defer">
res += 'c';
</script>
<script>
res += 'B';
</script>
<script>
res += 'C';

s = document.createElement('script');
s.src="file_bug28293.sjs?res+='d';";
s.defer = true;
document.body.appendChild(s);

s = document.createElement('script');
s.textContent="res+='D';";
document.body.appendChild(s);

res += 'E';
</script>
<script>
res += 'F';
document.addEventListener("DOMContentLoaded", function() {
  res += '1'
  s = document.createElement('script');
  s.src="file_bug28293.sjs?res+='M';";
  document.body.appendChild(s);
}, false);
res += 'G';
</script>
<script defer="defer">
res += 'e';
</script>
<script src="file_bug28293.sjs?res+='H';"></script>
<script>
<![CDATA[
res += 'I';
s = document.createElement('script');
s.src="file_bug28293.sjs?fHadExecuted=(res.indexOf('f')>=0);";
document.body.appendChild(s);
res += 'J';
]]>
</script>
<script defer="defer">
res += 'f';
</script>

</body>
</html>
